Remove console logs, update game action logic, move createSelectors out of game state logic into store index
