+++ src/stores/index.ts
@@ -1,0 +1,20 @@
+import { StoreApi, UseBoundStore } from 'zustand'
+import { gameStore } from './useGameStore'
+
+type WithSelectors<S> = S extends { getState: () => infer T }
+  ? S & { use: { [K in keyof T]: () => T[K] } }
+  : never
+
+const createSelectors = <S extends UseBoundStore<StoreApi<object>>>(
+  _store: S,
+) => {
+  const store = _store as WithSelectors<typeof _store>
+  store.use = {}
+  for (const k of Object.keys(store.getState())) {
+    (store.use as any)[k] = () => store((s) => s[k as keyof typeof s])
+  }
+
+  return store
+}
+
+export const useGameStateStore = createSelectors(gameStore)
